backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
transformers==4.36.0
torch==2.1.0
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
requests==2.31.0
aiofiles==23.2.1
ibm-watson==7.0.1
ibm-cloud-sdk-core==3.18.0
pydantic-settings==2.1.0
google-generativeai==0.3.2
sqlalchemy==2.0.23
alembic==1.13.1
